Change AGC parameters for SSB operation
This changes AGC to 'very slow' for SSB. It requires a change to pycsdr per the information therein (and re-run the ./setup.py install install_headers if making this change to an existing installation) and is designed to give an experience similar to a TS930s slow mode. This should mean a quieter receive in noisy condx. Changing 'AgcProfile.VSLOW' to AgcProfile.SLOW' will set the Ssb operation to slow agc, already present in an existing pycsdr instance.