JS smeter bar, sig performance improvement
CSS transform seems to need a lot of cpu on remote machines (using Linux + FF). Manageable if machine is reasonable but causes audio stutter and other issues on low-end devices. This implements a js meter bar that seems to need a lot less hp. Maybe CSS is ok on Win + other browser but...?