introduce improved spectrum display
Significant rewrite to the spectrum function originally incorporated into spectrum.js. This now uses data direct from waterfall_mkcolor to feed into the spectrum function, it is also called at the requisite time, and shut off when not needed. Better filtering is introduced (yet to determine which is preferable).

There is further tidying and efficiencies to be carried out, however it appears to be a good start for now.