Improvements to spectrum display code
Tidies up the code a little more and moves much of the init stuff to a new function. This remains an interim step as there's more to do.

Still using observe for the present, which will result in non-fatal error if the spectrum window is closed after use, but it works for now. Needs to use the existing resize facility within the mature code at some point.