Add client info over waterfall
Puts a small blue box (over the waterfall) containing the number of users connected to the system. Removes the 'under construction' box as this seems of little real utility. Also minor change to title identifying this version of the software and referring to general UI changes that will appear here shortly.